debian:  Added a quilt Debain package to build some required xen packages
Note that the Debian section is actually 2 levels deep here, under the
xen-4.4.1 directory since quilt need the tar.gz to be untared which would
have been too messy in the root directory.  builder Makefiles have been
adjusted to hadle this.

There are no 4.4 releases for Debian so a new package needed to be created
which provides:
- Package: libxen-4.4-qubes-vm
- Package: libxenstore3.0
- Package: libxen-qubes-vm-dev