Use proper make dependencies for verify-sources
This prevents the commands from re-running when we update
sources and the files havent changed. Since there is now a
full dependency chain there is no need to have a separate
verify-sources step either, in the future the build target(s)
can simply declare signature checks as a prerequisite.

This latter change is not made yet since get-sources from
qubes-builder expects the target to exist (for now).