stubdom-linux: rework pci add/remove handling
Use qemu_pci_{add,remove}_xenstore as used for minios stubdoms instead
of triggering on new PCI devices inside the stubdom. This way we have
the same timing behavior as with minios stubdoms or qemu in dom0. As a
bonus this makes the patch smaller.

BTW: Fix logic in remove case.