libxl: always use qemu format=host_device for devices hosted by stubdomain
Original target path is not available there. In case of device being
hosted in non-dom0, it isn't available in dom0 either, so don't try
accessing it there either.

Also, for cdrom, don't force specific device index. Other disks are
provided as SCSI, which are counted independently. And IDE interface is
limited to maximum index of 4.

QubesOS/qubes-issues#2951