rpm: move libvirtd restart to posttrans
Do not restart libvirtd in the middle of rpm transaction. This might
happen when only some of libvirt packages are upgraded (there is strict
dependency between libvirtd-daemon and libvirtd-libs). If libvirtd fails
to restart, then %posttrans of libvirtd package would not start it
either (because it wasn't running at that point).