Park unused CPU threads after suspend-to-RAM
This commit introduces nine commits that let Xen properly "park" unused
CPU threads when resuming from suspend-to-RAM state.

Without this commit and with "smt=0" on Xen's command line, my laptop's
CPU consumes more power after resuming from a suspend-to-RAM operation,
as measured with powertop 2.13 in dom0, compared to the case with this
commit.

Below are the averages of power measurements before/after suspend-to-RAM
(S2R) operations without and with this patch, obtained with the command
"sudo powertop --time=30 --csv=log.csv" invoked before and after the
S2R operation. In particular, powertop's estimation of the "system
baseline power" was used as the power consumption metric. Please note
that the CPU cores were forced to run at their lowest frequencies with
"xenpm" before and after the suspend (as the setting appears to be lost
during the suspend operation), and cron was stopped after every boot-up,
to ensure valid comparisons.

* Without this commit, after boot-up, but before S2R:
  874 mW (average of 24 measurements in 8 experiment iterations, where
          each iteration resulted in 3 measurements)

* Without this commit, after resuming from S2R:
  2.26 W (average of 24 measurements in 8 experiment iterations, with
          3 measurements per iteration)

* With this commit, after boot-up, but before S2R:
  826 mW (average of 27 measurements in 9 experiment iterations, with
          3 measurements per iteration)

* With this commit, after resuming from S2R:
  1.16 W (average of 27 measurements in 9 experiment iterations, with
          3 measurements per iteration)

As can be observed from the results, the reduction of power consumption
after resuming from S2R is significant when the cases without and with
this commit are compared: 2.26 W vs. 1.16 W.

Please note that I have not been able to explain why there is still
a difference before and after the S2R operation with this commit, albeit
a smaller one: 826 mW vs. 1.16 W.

Also note that, of the 9 new patches introduced by this commit, the
last patch resolves intermittent segmentation faults encountered by
systemd-sleep in dom0 during the resume from S2R when the remaining
patches are applied. The segmentation faults occur because the %fs
register is set by
  xen/arch/x86/boot/wakeup.S
during the resume from S2R, which, due to assumptions in the context
switching code in Xen and Linux, leaks to systemd-sleep running in
dom0 user-space, and systemd-sleep crashes with a segmentation fault
when it attempts to access a variable in thread-local storage. For
further details, please see the following links:
* https://github.com/QubesOS/qubes-issues/issues/5210
* https://lists.xen.org/archives/html/xen-devel/2020-07/msg00564.html
* https://lists.xen.org/archives/html/xen-devel/2020-07/msg00937.html