libxl: correctly detect backend_domid when detaching a device
libvirt does not set backend_domid when detaching a disk device (it does
backend_name, but it isn't resolved on detach). Instead of resolving the
backend_name, use other information (that is sufficient already) to
deduce backend_domid.

The patch is upstreamable, but likely only after Xen 4.16 will be
released (as in: after the pre-release code freeze).

Fixes QubesOS/qubes-issues#6996