Two more places to extend IGD check
The VGA class is checked in two places. The second one is related to
allowing opregion access - needed to access VBT.

And the second place is in hvmloader, which need to actually reserve
area for the opregion in the memory map - here also lookup for IGD need
device class adjustment.

QubesOS/qubes-issues#8561